Move GameMode into a new enums namespace
